fixed minior lapsus in Font (alex, you had one job...)

+++ Source/Framework/OpenGL/Font.cpp
@@ -159,7 +159,7 @@ void Font::setColor(const Color& c)
 	glUseProgram(0);
 }
 
-int Font::getMaxHeight() const
+float Font::getMaxHeight() const
 {
 	return m_fontHeight;
 }
